feat: replace "memory_stats" with peak memory measure in zkvms
